BACKPORT: soundwire: bus: Make sdw_nwrite() data pointer argument const
Idiomatically, write functions should take const pointers to the
data buffer, as they don't change the data. They are also likely
to be called from functions that receive a const data pointer.

Internally the pointer is passed to function/structs shared with
the read functions, requiring a cast, but this is an implementation
detail that should be hidden by the public API.

Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>
Link: https://lore.kernel.org/r/20210616145901.29402-1-rf@opensource.cirrus.com
Signed-off-by: Vinod Koul <vkoul@kernel.org>

BUG=b:238403794
TEST=Test Audio use cases

 Conflicts:
	include/linux/soundwire/sdw.h
(no code changes)
Signed-off-by: sboyane <shankarx.boyane@intel.com>
Change-Id: I213301ddf58bf1cc1606d9fc0db2ffff5a64d158